Revert "feat: Add a Linked accounts settings page"
This reverts commit afde53c97bce721d6287bf2e7e059fd01dff6744.
