+++ web_src/src/App.tsx
@@ -39,7 +39,6 @@ import {
   FactorySettingsLayout,
   FactorySettingsAccountNotificationsPage,
   FactorySettingsAccountProfilePage,
-  FactorySettingsAccountLinkedAccountsPage,
   FactorySettingsAccountSecurityPage,
   FactorySettingsRepositoryPage,
   FactorySettingsUsagePage,
@@ -306,11 +305,6 @@ const factorySettingsSectionRoutes = [
     key="factory-settings-account-profile"
     path="account/profile"
     element={<FactorySettingsAccountProfilePage />}
-  />,
-  <Route
-    key="factory-settings-account-linked-accounts"
-    path="account/linked-accounts"
-    element={<FactorySettingsAccountLinkedAccountsPage />}
   />,
   <Route
     key="factory-settings-account-security"
