+++ web_src/src/pages/factories/index.ts
@@ -12,7 +12,6 @@ export {
   FactorySettingsLayout,
   FactorySettingsAccountNotificationsPage,
   FactorySettingsAccountProfilePage,
-  FactorySettingsAccountLinkedAccountsPage,
   FactorySettingsAccountSecurityPage,
   FactorySettingsNotificationsPage,
   FactorySettingsProfilePage,
