+++ web_src/src/pages/factories/pages/settings/account-profile-redesign/AccountProfileRedesignPlayground.tsx
@@ -8,7 +8,6 @@ import {
 } from "./accountProfileRedesignMocks";
 import { AccountProfileRedesignShell } from "./AccountProfileRedesignShell";
 import {
-  AccountLinkedAccountsRedesignRoutePage,
   AccountProfileRedesignProvider,
   AccountProfileRedesignRoutePage,
   AccountSecurityRedesignRoutePage,
@@ -37,7 +36,6 @@ export function AccountProfileRedesignPlayground({
         onSelectPage={setPage}
       >
         {page === "profile" ? <AccountProfileRedesignRoutePage /> : null}
-        {page === "linked-accounts" ? <AccountLinkedAccountsRedesignRoutePage /> : null}
         {page === "security" ? <AccountSecurityRedesignRoutePage /> : null}
       </AccountProfileRedesignShell>
       <Toaster position="bottom-center" closeButton />
