+++ web_src/src/pages/factories/pages/settings/account-profile-redesign/accountProfileRedesignMocks.ts
@@ -1,6 +1,6 @@
 import { defaultNotificationTypeToggles, type NotificationTypeToggles } from "@/lib/notificationSettings";
 
-export type AccountRedesignPageId = "profile" | "linked-accounts" | "security";
+export type AccountRedesignPageId = "profile" | "security";
 
 export type AccountRedesignWorkspaceScope = "all" | "selected";
