+++ web_src/src/pages/factories/pages/settings/settingsNavItems.spec.ts
@@ -5,9 +5,6 @@ import { FACTORY_SETTINGS_NAV_GROUPS, factorySettingsRouteFromPathname } from ".
 describe("factorySettingsRouteFromPathname", () => {
   it("reads a canonical scoped settings route", () => {
     expect(factorySettingsRouteFromPathname("/org/workspaces/RF/settings/account/profile")?.id).toBe("account-profile");
-    expect(factorySettingsRouteFromPathname("/org/workspaces/RF/settings/account/linked-accounts")?.id).toBe(
-      "account-linked-accounts",
-    );
     expect(factorySettingsRouteFromPathname("/org/workspaces/RF/settings/workspace/automations")?.id).toBe(
       "workspace-automations",
     );
@@ -27,7 +24,6 @@ describe("FACTORY_SETTINGS_NAV_GROUPS", () => {
     expect(FACTORY_SETTINGS_NAV_GROUPS.map((group) => group.label)).toEqual(["Account", "Workspace", "Organization"]);
     expect(FACTORY_SETTINGS_NAV_GROUPS.flatMap((group) => group.items.map((item) => item.label))).toEqual([
       "Profile",
-      "Linked accounts",
       "Security",
       "Notifications",
       "General",
