+++ web_src/src/pages/factories/pages/settings/settingsNavItems.ts
@@ -8,7 +8,6 @@ import {
   Grid3x3,
   Key,
   KeyRound,
-  Link2,
   Plug,
   Settings,
   Shield,
@@ -21,7 +20,6 @@ import type { FactorySettingsScope } from "../../lib/factoryPagePaths";
 export type FactorySettingsSection =
   | "general"
   | "profile"
-  | "linked-accounts"
   | "security"
   | "notifications"
   | "repository"
@@ -53,13 +51,6 @@ export const FACTORY_SETTINGS_NAV_GROUPS: FactorySettingsNavGroup[] = [
     label: "Account",
     items: [
       { id: "account-profile", label: "Profile", Icon: CircleUser, scope: "account", section: "profile" },
-      {
-        id: "account-linked-accounts",
-        label: "Linked accounts",
-        Icon: Link2,
-        scope: "account",
-        section: "linked-accounts",
-      },
       { id: "account-security", label: "Security", Icon: Shield, scope: "account", section: "security" },
       { id: "account-notifications", label: "Notifications", Icon: Bell, scope: "account", section: "notifications" },
     ],
